requirements.txt
Logbook==0.3
distribute==0.6.19
httplib2==0.7.2
oauth2==1.5.211
python-twitter==0.8.2
simplejson==2.3.3
wsgiref==0.1.2
